% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuracion_diapositiva.R
\name{configuracion_diapositiva}
\alias{configuracion_diapositiva}
\title{Agregar propiedades a la diapositiva actual}
\usage{
configuracion_diapositiva(., localizacion, enunciado, propiedades, alineacion)
}
\description{
Agregar propiedades a la diapositiva actual
}
