% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuadr_grafico.R
\name{cuadro_grafico}
\alias{cuadro_grafico}
\title{Agregar propiedades a la diapositiva actual}
\usage{
cuadro_grafico(., localizacion, grafico)
}
\description{
Agregar propiedades a la diapositiva actual
}
