% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuadro_texto.R
\name{cuadro_texto}
\alias{cuadro_texto}
\title{Agregar propiedades a la diapositiva actual, nos enfocamos en un cuadro con texto}
\usage{
cuadro_texto(., localizacion, enunciado, propiedades, alineacion)
}
\description{
Agregar propiedades a la diapositiva actual, nos enfocamos en un cuadro con texto
}
