% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{EpiPowerPoint es un paquete libre para diseño de presentaciones desde el entorno de RStudio}
\usage{
hello()
}
\description{
EpiPowerPoint es un paquete libre para diseño de presentaciones desde el entorno de RStudio
}
