% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_pptx.R
\name{leer_pptx}
\alias{leer_pptx}
\title{EpiPowerPoint es un paquete libre para diseño de presentaciones desde el entorno de RStudio}
\usage{
leer_pptx(direccion = "bin/pptx/Plantilla_2022.pptx")
}
\description{
EpiPowerPoint es un paquete libre para diseño de presentaciones desde el entorno de RStudio
}
